Migrate to Gradle Kotlin DSL and update native libs
Replaced build.gradle with build.gradle.kts, updating the build system to use the Gradle Kotlin DSL and modernizing dependency management. Removed bundled native JARs for Processing/JOGL and added platform-specific native libraries as loose files under library/processing/core/library for multiple OS/architectures. Updated .idea and .vscode configs for improved IDE support, added new Processing net library and examples, and upgraded Gradle wrapper to 9.0.0.

+++ src/directions/subscene/generateTree/RTree.java
@@ -147,10 +147,10 @@ public class RTree {
                     Vector parentPos = n.getParent().getPos();
                     PShape lines;
                     float angle = PApplet.radians(36);
-                    float x = parentPos.x + RTreeNode.radius * PApplet.sin(angle) * (2 * n.getChildNumber() - 1);
-                    float y = parentPos.y + RTreeNode.radius * PApplet.cos(angle);
+                    double x = parentPos.x + RTreeNode.radius * PApplet.sin(angle) * (2 * n.getChildNumber() - 1);
+                    double y = parentPos.y + RTreeNode.radius * PApplet.cos(angle);
 
-                    lines = p.createShape(LINE, x, y, c * n.pos.x + (1 - c) * x, c * (n.pos.y - RTreeNode.radius) + (1 - c) * y);
+                    lines = p.createShape(LINE, (float) x, (float) y, (float) (c * n.pos.x + (1 - c) * x), (float) (c * (n.pos.y - RTreeNode.radius) + (1 - c) * y));
                     n.getParent().getNodeShape().addChild(lines);
                 } else
                     n.draw(c, skeleton);
